Reorganises transferts files
